Issue-9765 - Extract array_resize and cardinality functions from functions-array subcrate' s kernels and udf containers
