+++ datafusion/functions-array/src/kernels.rs
@@ -18,79 +18,18 @@
 //! implementation kernels for array functions
 
 use arrow::array::{
-    Array, ArrayRef, Capacities, GenericListArray, Int64Array, ListArray,
-    MutableArrayData, OffsetSizeTrait, UInt64Array,
+    Array, ArrayRef, GenericListArray, ListArray, OffsetSizeTrait, UInt64Array,
 };
 use arrow::datatypes::{DataType, UInt64Type};
-use arrow_buffer::{ArrowNativeType, OffsetBuffer};
-use arrow_schema::FieldRef;
+use arrow_buffer::OffsetBuffer;
 
 use datafusion_common::cast::{
-    as_generic_list_array, as_int64_array, as_large_list_array, as_list_array,
+    as_generic_list_array, as_large_list_array, as_list_array,
 };
-use datafusion_common::{
-    exec_err, internal_datafusion_err, DataFusionError, Result, ScalarValue,
-};
+use datafusion_common::{exec_err, Result};
 
-use crate::utils::downcast_arg;
-use std::any::type_name;
+use crate::utils::compute_array_dims;
 use std::sync::Arc;
-
-/// Returns the length of each array dimension
-fn compute_array_dims(arr: Option<ArrayRef>) -> Result<Option<Vec<Option<u64>>>> {
-    let mut value = match arr {
-        Some(arr) => arr,
-        None => return Ok(None),
-    };
-    if value.is_empty() {
-        return Ok(None);
-    }
-    let mut res = vec![Some(value.len() as u64)];
-
-    loop {
-        match value.data_type() {
-            DataType::List(..) => {
-                value = downcast_arg!(value, ListArray).value(0);
-                res.push(Some(value.len() as u64));
-            }
-            _ => return Ok(Some(res)),
-        }
-    }
-}
-
-fn generic_list_cardinality<O: OffsetSizeTrait>(
-    array: &GenericListArray<O>,
-) -> Result<ArrayRef> {
-    let result = array
-        .iter()
-        .map(|arr| match compute_array_dims(arr)? {
-            Some(vector) => Ok(Some(vector.iter().map(|x| x.unwrap()).product::<u64>())),
-            None => Ok(None),
-        })
-        .collect::<Result<UInt64Array>>()?;
-    Ok(Arc::new(result) as ArrayRef)
-}
-
-/// Cardinality SQL function
-pub fn cardinality(args: &[ArrayRef]) -> Result<ArrayRef> {
-    if args.len() != 1 {
-        return exec_err!("cardinality expects one argument");
-    }
-
-    match &args[0].data_type() {
-        DataType::List(_) => {
-            let list_array = as_list_array(&args[0])?;
-            generic_list_cardinality::<i32>(list_array)
-        }
-        DataType::LargeList(_) => {
-            let list_array = as_large_list_array(&args[0])?;
-            generic_list_cardinality::<i64>(list_array)
-        }
-        other => {
-            exec_err!("cardinality does not support type '{:?}'", other)
-        }
-    }
-}
 
 /// Array_dims SQL function
 pub fn array_dims(args: &[ArrayRef]) -> Result<ArrayRef> {
@@ -156,100 +95,6 @@ pub fn array_ndims(args: &[ArrayRef]) -> Result<ArrayRef> {
         }
         array_type => exec_err!("array_ndims does not support type {array_type:?}"),
     }
-}
-
-/// array_resize SQL function
-pub fn array_resize(arg: &[ArrayRef]) -> Result<ArrayRef> {
-    if arg.len() < 2 || arg.len() > 3 {
-        return exec_err!("array_resize needs two or three arguments");
-    }
-
-    let new_len = as_int64_array(&arg[1])?;
-    let new_element = if arg.len() == 3 {
-        Some(arg[2].clone())
-    } else {
-        None
-    };
-
-    match &arg[0].data_type() {
-        DataType::List(field) => {
-            let array = as_list_array(&arg[0])?;
-            general_list_resize::<i32>(array, new_len, field, new_element)
-        }
-        DataType::LargeList(field) => {
-            let array = as_large_list_array(&arg[0])?;
-            general_list_resize::<i64>(array, new_len, field, new_element)
-        }
-        array_type => exec_err!("array_resize does not support type '{array_type:?}'."),
-    }
-}
-
-/// array_resize keep the original array and append the default element to the end
-fn general_list_resize<O: OffsetSizeTrait>(
-    array: &GenericListArray<O>,
-    count_array: &Int64Array,
-    field: &FieldRef,
-    default_element: Option<ArrayRef>,
-) -> Result<ArrayRef>
-where
-    O: TryInto<i64>,
-{
-    let data_type = array.value_type();
-
-    let values = array.values();
-    let original_data = values.to_data();
-
-    // create default element array
-    let default_element = if let Some(default_element) = default_element {
-        default_element
-    } else {
-        let null_scalar = ScalarValue::try_from(&data_type)?;
-        null_scalar.to_array_of_size(original_data.len())?
-    };
-    let default_value_data = default_element.to_data();
-
-    // create a mutable array to store the original data
-    let capacity = Capacities::Array(original_data.len() + default_value_data.len());
-    let mut offsets = vec![O::usize_as(0)];
-    let mut mutable = MutableArrayData::with_capacities(
-        vec![&original_data, &default_value_data],
-        false,
-        capacity,
-    );
-
-    for (row_index, offset_window) in array.offsets().windows(2).enumerate() {
-        let count = count_array.value(row_index).to_usize().ok_or_else(|| {
-            internal_datafusion_err!("array_resize: failed to convert size to usize")
-        })?;
-        let count = O::usize_as(count);
-        let start = offset_window[0];
-        if start + count > offset_window[1] {
-            let extra_count =
-                (start + count - offset_window[1]).try_into().map_err(|_| {
-                    internal_datafusion_err!(
-                        "array_resize: failed to convert size to i64"
-                    )
-                })?;
-            let end = offset_window[1];
-            mutable.extend(0, (start).to_usize().unwrap(), (end).to_usize().unwrap());
-            // append default element
-            for _ in 0..extra_count {
-                mutable.extend(1, row_index, row_index + 1);
-            }
-        } else {
-            let end = start + count;
-            mutable.extend(0, (start).to_usize().unwrap(), (end).to_usize().unwrap());
-        };
-        offsets.push(offsets[row_index] + count);
-    }
-
-    let data = mutable.freeze();
-    Ok(Arc::new(GenericListArray::<O>::try_new(
-        field.clone(),
-        OffsetBuffer::<O>::new(offsets.into()),
-        arrow_array::make_array(data),
-        None,
-    )?))
 }
 
 // Create new offsets that are euqiavlent to `flatten` the array.
